Flytter logikk til egen komponent
